+++ sqlite3_web/web/controller.dart
@@ -12,7 +12,7 @@ final class ExampleController extends DatabaseController {
   @override
   Future<JSAny?> handleCustomRequest(
     ClientConnection connection,
-    JSAny? request,
+    CustomClientRequest request,
   ) async {
     return null;
   }
@@ -49,7 +49,7 @@ final class ExampleDatabase extends WorkerDatabase {
   @override
   Future<JSAny?> handleCustomRequest(
     ClientConnection connection,
-    JSAny? request,
+    CustomClientRequest request,
   ) async {
     final response =
         ((await connection.customRequest(null)) as JSNumber).toDartInt;