+++ sqlite3_web/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.7.0-wip
+
+- __Breaking__: Introduce `CustomClientRequest` and `CustomClientDatabaseRequest` classes representing
+  client requests. They expose an abort signal for requests and a way to interact with database locks.
+- Add `LockToken? token` and `Future<void>? abortTrigger` parameters to `Database.customRequest`.
+
 ## 0.6.0
 
 - Replace `worker` URI parameter with `WorkerConnector` interface. Use